[FIX] purchase_delivery_split_date
When the PO is already confirmed and you add a new line with a different date than the existing one, the new receipt is created but the old receipt and move have the dates updated with no reason. The call on the method _update_picking_from_group_key is the method that is messing it up as it updates the previous picking. Also ensures that date is updated in the stock.moves when the value changes in purchase.order.line